Fixes #2751

Changes made in this Pull Request:
 - AtomNames now use a lookup table in their implementation


PR Checklist
------------
 - [x] Tests?
 - [ ] Docs?
 - [x] CHANGELOG updated?
 - [x] Issue raised/referenced?