@richardjgowers 2020-06-14
This is a different way of storing atom names in the `Topology` object, where instead of an array of `['N', 'C', etc]` you instead have two arrays, one a set of known atom names, i.e. `['C', 'N', ...]` then the atom names are stored as indices into this array.

This is then beneficial for quickly matching strings (ie names), as you match against a **set** of names to yield integers, then quickly match integers. There's an assumption that the number of unique atom names is very small compared to the number of atoms in the system, ie atom names are repeated often.

I did some timing with this sequence of actions:
```python
import MDAnalysis as mda
from MDAnalysisTests.datafiles import GRO

u = mda.Universe(GRO)

names = u.atoms.names

ag = u.select_atoms('name O*')

ag.names = 'X' 
```

* Universe creation is a bit slower (600 -> 625 ms) for me.
* Attribute access is identical.
* `select_atoms` goes from 69 ms to 0.69 ms
* setting atom names gets slower from 0.6ms to 1.1ms.

I think I can probably make Universe creation faster by rewriting how the Parsers are building the AtomName attribute (ie build the lookup array as I go).  I think atoms are selected more often than they are set, so I'm not too concerned about the performance drop in attribute setting.

Then in a very hand wavy way, I think this will also help serialisation efforts as this is more compact to store/send.