@richardjgowers 2020-06-14
So if @MDAnalysis/coredevs could try this on their typical workloads and see if it works for them.