@richardjgowers 2020-06-20
@IAlibay as a way of slowing you down from making more PRs, can you review this?  Is it understandable/maintainable and does it show performance improvements on your typical systems?