@richardjgowers 2020-06-21
Ok I've made all string topology attrs use this approach.  I think I can improve selections like `protein` to leverage the new stuff too.