@IAlibay 2020-06-21
So the code adds a numpy array and a smaller (unless everything is unique) list + dictionary set per attribute. At what point do we need to consider the extra memory cost of Universe creation?