@richardjgowers 2020-06-21
The memory cost is likely almost always smaller. Yes there are more python
objects, but each resname is stored as an integer (2 bytes for uint16)
rather than a Python string, likely 3+? I’ll double check these assumptions
though :)

On Sun, Jun 21, 2020 at 21:27, Irfan Alibay <notifications@github.com>
wrote:

> So the code adds a numpy array and a smaller (unless everything is unique)
> list + dictionary set per attribute. At what point do we need to consider
> the extra memory cost of Universe creation?
>
> —
> You are receiving this because you authored the thread.
> Reply to this email directly, view it on GitHub
> <https://github.com/MDAnalysis/mdanalysis/pull/2755#issuecomment-647177527>,
> or unsubscribe
> <https://github.com/notifications/unsubscribe-auth/ACGSGB7MAAIAJJL2RKSFJRTRXZUKNANCNFSM4N5OMNXQ>
> .
>