@IAlibay 2020-06-22
Update on my million atom benchmark for this new patch.

  - Universe creation has further increased to 12.4s from 10.9s (develop).
  - Selecting waters has decreased to 54.2 ms from 1.98s (develop)
  - water renaming hasn't substantially changed (slight reduction but within error).

It's probably worth noting that I suspect that the TOPParser is particularly inefficient, so it's likely that the difference in universe creation is going to be a lot different if say a GRO file was read.