@richardjgowers 2020-06-22
@IAlibay thanks for benchmarking.  I'm not shocked that this is a little slower to construct, we're constructing an index of possible types, so there's a Python loop over all atom values, for each attribute.  I'll see if I can take the edge off that... Killing a 2s selection time is nice though :)