@orbeckst 2020-06-30
Is the only thing left here the refactoring of the duplicated code? It would be nice to get this regression fixed sooner than later (also for PR #2798 ).