@pep8speaks 2020-07-05
Hello @richardjgowers! Thanks for updating this PR. We checked the lines you've touched for [PEP 8](https://www.python.org/dev/peps/pep-0008) issues, and found:

* In the file [`package/MDAnalysis/core/selection.py`](https://github.com/MDAnalysis/mdanalysis/blob/00ba0ee77d43f7b799158994445182aabd4f841c/package/MDAnalysis/core/selection.py):

> [Line 577:10](https://github.com/MDAnalysis/mdanalysis/blob/00ba0ee77d43f7b799158994445182aabd4f841c/package/MDAnalysis/core/selection.py#L577): [E225](https://duckduckgo.com/?q=pep8%20E225) missing whitespace around operator

* In the file [`package/MDAnalysis/core/topologyattrs.py`](https://github.com/MDAnalysis/mdanalysis/blob/00ba0ee77d43f7b799158994445182aabd4f841c/package/MDAnalysis/core/topologyattrs.py):

> [Line 479:1](https://github.com/MDAnalysis/mdanalysis/blob/00ba0ee77d43f7b799158994445182aabd4f841c/package/MDAnalysis/core/topologyattrs.py#L479): [W293](https://duckduckgo.com/?q=pep8%20W293) blank line contains whitespace
> [Line 527:80](https://github.com/MDAnalysis/mdanalysis/blob/00ba0ee77d43f7b799158994445182aabd4f841c/package/MDAnalysis/core/topologyattrs.py#L527): [E501](https://duckduckgo.com/?q=pep8%20E501) line too long (83 > 79 characters)
> [Line 1781:1](https://github.com/MDAnalysis/mdanalysis/blob/00ba0ee77d43f7b799158994445182aabd4f841c/package/MDAnalysis/core/topologyattrs.py#L1781): [W293](https://duckduckgo.com/?q=pep8%20W293) blank line contains whitespace
> [Line 1800:51](https://github.com/MDAnalysis/mdanalysis/blob/00ba0ee77d43f7b799158994445182aabd4f841c/package/MDAnalysis/core/topologyattrs.py#L1800): [W291](https://duckduckgo.com/?q=pep8%20W291) trailing whitespace
> [Line 1829:80](https://github.com/MDAnalysis/mdanalysis/blob/00ba0ee77d43f7b799158994445182aabd4f841c/package/MDAnalysis/core/topologyattrs.py#L1829): [E501](https://duckduckgo.com/?q=pep8%20E501) line too long (83 > 79 characters)
> [Line 1832:51](https://github.com/MDAnalysis/mdanalysis/blob/00ba0ee77d43f7b799158994445182aabd4f841c/package/MDAnalysis/core/topologyattrs.py#L1832): [W291](https://duckduckgo.com/?q=pep8%20W291) trailing whitespace
> [Line 2022:1](https://github.com/MDAnalysis/mdanalysis/blob/00ba0ee77d43f7b799158994445182aabd4f841c/package/MDAnalysis/core/topologyattrs.py#L2022): [W293](https://duckduckgo.com/?q=pep8%20W293) blank line contains whitespace
> [Line 2041:51](https://github.com/MDAnalysis/mdanalysis/blob/00ba0ee77d43f7b799158994445182aabd4f841c/package/MDAnalysis/core/topologyattrs.py#L2041): [W291](https://duckduckgo.com/?q=pep8%20W291) trailing whitespace
> [Line 2070:80](https://github.com/MDAnalysis/mdanalysis/blob/00ba0ee77d43f7b799158994445182aabd4f841c/package/MDAnalysis/core/topologyattrs.py#L2070): [E501](https://duckduckgo.com/?q=pep8%20E501) line too long (83 > 79 characters)
> [Line 2073:51](https://github.com/MDAnalysis/mdanalysis/blob/00ba0ee77d43f7b799158994445182aabd4f841c/package/MDAnalysis/core/topologyattrs.py#L2073): [W291](https://duckduckgo.com/?q=pep8%20W291) trailing whitespace
> [Line 2075:1](https://github.com/MDAnalysis/mdanalysis/blob/00ba0ee77d43f7b799158994445182aabd4f841c/package/MDAnalysis/core/topologyattrs.py#L2075): [W293](https://duckduckgo.com/?q=pep8%20W293) blank line contains whitespace