@orbeckst 2020-07-13
AppVeyor fails on *TestICodes.test_set_residues_singular* – is this related to this PR?