@richardjgowers 2020-07-28
@MDAnalysis/coredevs does someone fancy giving this a once over?