@orbeckst 2020-08-16
@IAlibay @jbarnoud do you see any blocking issues here?

I am keen to see if it looks better in our benchmarks.