@jbarnoud 2020-08-16
I am not in a position to review in depth, thought from what I read on my phone it looks good to me. I am uneasy with the amount of code duplication, but it can be a separate pull request.